Add missing files
